3 x x + 2 x x ruins MITM mitigation
When you are not your own intermediary then anyone can issue (for free or buy) a valid certificate. Effectively ruining MITM mitigation.

The rollover info was over complex and exposes the added security layer of DANE. Unless you would be that root authority and issuing certificates for private use only (which is highly unlikely).